// packages/sirius-web/frontend/sirius-web-application/src/modals/update-library/UpdateLibraryTable.types.ts
/*******************************************************************************
 * Copyright (c) 2025 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/
import { GQLGetLibraryVersionsQueryData } from './useLibraryVersions.types';

export interface UpdateLibraryTableProps {
  namespace: string;
  name: string;
  version: string;
  onSelectionChange(selection: string[]);
}

export interface UpdateLibraryTableState {
  data: GQLGetLibraryVersionsQueryData | null;
}